+++ app/Widgets/ContactActions/ContactActions.php
@@ -8,6 +8,7 @@ use App\Widgets\Dialog\Dialog;
 use App\Widgets\Drawer\Drawer;
 use App\Widgets\Post\Post;
 use App\Widgets\Toast\Toast;
+use Movim\CurrentCall;
 use Movim\EmbedLight;
 use Movim\Widget\Base;
 
@@ -75,6 +76,7 @@ class ContactActions extends Base
         $hasFingerprints = ($this->user->bundles()->where('jid', $jid)->count() > 0);
 
         $tpl->assign('jid', $jid);
+        $tpl->assign('incall', CurrentCall::getInstance()->isStarted());
         $tpl->assign('clienttype', getClientTypes());
         $tpl->assign('hasfingerprints', $hasFingerprints);
         $tpl->assign('posts', \App\Post::where('server', $jid)
@@ -139,7 +141,7 @@ class ContactActions extends Base
         $c = new Chats();
         $c->ajaxOpen($jid);
 
-        $this->rpc('MovimUtils.redirect', $this->route('chat', $jid));
+        $this->rpc('MovimUtils.reload', $this->route('chat', $jid));
     }
 
     public function ajaxHttpGetPictures($jid, $page = 0)